[Quant][core][devs] Separated implementations for quantized & non-quantized tensors in reflection_pad2d_cpu
Summary: This PR is part of a series of PRs addressing https://github.com/pytorch/pytorch/issues/54150,
related to using dispatcher for calls to quantized backends as opposed to if/else conditionals.
This particular PR separates the calls to quantized & non-quantized backends for reflection_pad2d_cpu
using a dispatcher.

ghstack-source-id: 5015f48dda0a191ac2766cc4d1af44bbdb0c8ad5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/72442